+++ src/popup/components/Views/GroupView/GroupView.vue
@@ -51,6 +51,6 @@ watchEffect(() => {
             <Links v-else :group />
         </div>
 
-        <Message v-else> 😢 {{ trans('error_no_group_selected') }} </Message>
+        <Message v-else>😢 {{ trans('error_no_group_selected') }}</Message>
     </View>
 </template>