+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v1.6.0 (2024-12-01)
 - ✨ Added 1 more icon for groups
+- ✨ When the group is opened, it will show the number of tabs in the group
 
 ## v1.5.1 (2024-11-30)
 - 🧑‍💻 Change internalization to match [Chrome i18n API](https://developer.chrome.com/docs/extensions/reference/api/i18n#concepts_and_usage)